chore: improve types from config
