+++ packages/server/lib/project-base.ts
@@ -10,7 +10,7 @@ import { ServerCt, SocketCt } from '@packages/server-ct'
 import { SocketE2E } from './socket-e2e'
 import api from './api'
 import { Automation } from './automation'
-import config from './config'
+import * as config from './config'
 import cwd from './cwd'
 import errors from './errors'
 import Reporter from './reporter'
