+++ packages/server/test/unit/coerce_spec.js
@@ -1,7 +1,7 @@
 require('../spec_helper')
 
 const coerce = require(`${root}lib/util/coerce`)
-const getProcessEnvVars = require(`${root}lib/config`).getProcessEnvVars
+const getProcessEnvVars = require(`${root}lib/util/config`).getProcessEnvVars
 
 describe('lib/util/coerce', () => {
   beforeEach(function () {
